Fix OpenFlow version negotiation
The maximum version reported by the switch was always being used,
as a result if the libfluid application did not support this
version the negotiation would fail.

E.g. A switch supporting 1.0,1.3 and a controller supporting only 1.0
would fail rather than selecting version 1.0.

Now the highest version supported by both switch and controller is
selected as per the OpenFlow standard.